+++ System/Directory.hs
@@ -994,14 +994,17 @@ copyFileTimesFromStatus st dst = do
 -- current directory.  The function drops trailing path separators where
 -- possible (via 'dropTrailingPathSeparator').
 --
--- /Known bug(s)/: on Windows, the function does not resolve symbolic links
--- and the letter case of filenames is not canonicalized.
+-- /Known bugs/: When the path contains an existing symbolic link, but the
+-- target of the link does not exist, then the path is not dereferenced (bug
+-- #64).  On Windows, the function does not resolve symbolic links and the
+-- letter case of filenames is not canonicalized.
 --
 -- /Changes since 1.2.3.0:/ The function has been altered to be more robust
 -- and has the same exception behavior as 'makeAbsolute'.
 --
 -- /Changes since 1.3.0.0:/ The function no longer preserves the trailing path
--- separator.
+-- separator.  File symbolic links that appear in the middle of a path are
+-- properly dereferenced.
 --
 canonicalizePath :: FilePath -> IO FilePath
 canonicalizePath = \ path ->
@@ -1018,7 +1021,7 @@ canonicalizePath = \ path ->
     transform path = do
       encoding <- getFileSystemEncoding
       realpathPrefix encoding (reverse (zip prefixes suffixes)) path
-      where segments = splitPath path
+      where segments = splitDirectories path
             prefixes = scanl1 (</>) segments
             suffixes = tail (scanr (</>) "" segments)